backend-python/requirements.txt
Flask
mysql-connector-python
google-cloud-sqlcommenter
google-cloud-bigtable
google-cloud-core
mysql-connector-python
